test: move and add clean
